+++ simtools/params.py
@@ -1,9 +1,13 @@
 # -*- coding: utf-8 -*-
 """Parameter services.
 
-Parameter services provide loading parameters from a Python file.
+Parameter services provide the following functionality:
+
+- loading parameters from a Python file;
+- saving parameters to a JSON file.
 """
 
+import json
 import sys
 
 from simtools.base import Dict
@@ -33,6 +37,46 @@ class Params(Dict):
                 raise ParamFileError(filename=filename, lineno=lineno,
                                      error_msg=exc_value.args[0])
             self.update(new_params)
+
+    def save(self, filename, save_params=None, **kwargs):
+        """Save parameters to a file."""
+        DEFAULT_INDENT = 4
+
+        # If necessary, validate parameters to be saved
+        if save_params is not None:
+            try:
+                iter(save_params)
+            except TypeError:
+                raise TypeError("'save_params' is not iterable.")
+            try:
+                basestring
+            except NameError:
+                basestring = str
+            if isinstance(save_params, basestring):
+                raise TypeError("'save_params' is a string.")
+
+        # Validate extra keyword arguments
+        for arg in ('obj', 'fp'):
+            if arg in kwargs:
+                raise TypeError("save() got an unexpected keyword argument "
+                                "'{}'.".format(arg))
+
+        # Determine parameters to be saved
+        if save_params is not None:
+            try:
+                params = {p: self[p] for p in save_params}
+            except KeyError as e:
+                raise ValueError(
+                    "Selected parameter '{}' is not found.".format(e.args[0]))
+        else:
+            params = self
+
+        # Determine indentation
+        indent = kwargs.pop('indent', DEFAULT_INDENT)
+
+        # Save parameters to a JSON file
+        with open(filename, 'w') as params_file:
+            json.dump(params, params_file, indent=indent, **kwargs)
 
 
 def load_params(filename):