+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added parsing command line arguments.
 - Added loading parameters from a Python file.
+- Added saving parameters to a JSON file.
 - Added hierarchy of SimTools exceptions.
 - Added package installer.
 - Added package description.